requirements.txt
google-genai
Pillow
PyYAML
python-dotenv
grpcio
protobuf
